+++ pyticketswitch/__init__.py
@@ -1,4 +1,4 @@
 from pyticketswitch.client import Client  # NOQA
 
 
-__version__ = "2.8.7"
+__version__ = "2.9.1"