+++ tests/test_bundle.py
@@ -16,6 +16,9 @@ class TestBundle(object):
             'bundle_total_seatprice': 51,
             'bundle_total_send_cost': 1.5,
             'bundle_total_surcharge': 2.2,
+            'bundle_send_cost_tax_component': 0.95,
+            'bundle_total_combined_tax_component': 35.0,
+            'bundle_total_surcharge_tax_sub_component': 0.0,
             'currency_code': 'gbp',
             'order': [
                 {'item_number': 1},
@@ -52,12 +55,18 @@ class TestBundle(object):
 
         assert isinstance(bundle.total, (int, float))
         assert bundle.total == 54.7
+        
+        assert isinstance(bundle.send_cost_tax_component, (int, float))
+        assert bundle.send_cost_tax_component == 0.95
+
+        assert isinstance(bundle.total_combined_tax_component, (int, float))
+        assert bundle.total_combined_tax_component == 35.0
+
+        assert isinstance(bundle.total_surcharge_tax_sub_component, (int, float))
+        assert bundle.total_surcharge_tax_sub_component == 0.0
 
         assert isinstance(bundle.debitor, Debitor)
         assert bundle.debitor.type == 'dummy'
-
-        assert isinstance(bundle.purchase_result, PurchaseResult)
-        assert bundle.purchase_result.success
 
     def test_from_api_data_with_decimals(self):
         data = {
@@ -68,6 +77,9 @@ class TestBundle(object):
             'bundle_total_seatprice': Decimal('51'),
             'bundle_total_send_cost': Decimal('1.5'),
             'bundle_total_surcharge': Decimal('2.2'),
+            "bundle_send_cost_tax_component": Decimal('0.95'),
+            'bundle_total_combined_tax_component': Decimal('35.0'),
+            'bundle_total_surcharge_tax_sub_component': Decimal('0.0'),
             'currency_code': 'gbp',
             'order': [
                 {'item_number': 1},
@@ -105,11 +117,17 @@ class TestBundle(object):
         assert isinstance(bundle.total, Decimal)
         assert bundle.total == Decimal('54.7')
 
+        assert isinstance(bundle.send_cost_tax_component, Decimal)
+        assert bundle.send_cost_tax_component == Decimal('0.95')
+
+        assert isinstance(bundle.total_combined_tax_component, Decimal)
+        assert bundle.total_combined_tax_component == Decimal('35.0')
+
+        assert isinstance(bundle.total_surcharge_tax_sub_component, Decimal)
+        assert bundle.total_surcharge_tax_sub_component == Decimal('0.0')
+
         assert isinstance(bundle.debitor, Debitor)
         assert bundle.debitor.type == 'dummy'
-
-        assert isinstance(bundle.purchase_result, PurchaseResult)
-        assert bundle.purchase_result.success
 
     def test_get_events(self):